ci: [#64729] remove coverage
This commit removes coverage as the code is being tested by `pulumi`
however because it is being run by the `pulumi` runtime in another
process it does not seem that `coverage` can detect which lines are
being run.

It is possible coverage may be reintroduced in the future, however this
will probably require wrapping the `pulumi` runtime somehow.
The most likely candidate on how to do this is to run `coverage` around
the `pulumi` command by override the `pulumi` executable on the
`LocalWorkspaceOptions` argument to `create_stack` within the
`stack_builder` fixture, however it is deemed out-of-scope for now.